feat : 필요없는 DAO 삭제
