--- src/main/java/chess/dao/ConnectionSetup.java
@@ -1,34 +1,0 @@
-package chess.dao;
-
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-
-public class ConnectionSetup {
-    public Connection getConnection() {
-        Connection con = null;
-        String server = "localhost:3306"; // MySQL 서버 주소
-        String database = "study"; // MySQL DATABASE 이름
-        String option = "?useSSL=false&serverTimezone=UTC";
-        String userName = "jaeseongdev"; //  MySQL 서버 아이디
-        String password = "password"; // MySQL 서버 비밀번호
-
-        // 드라이버 로딩
-        try {
-            Class.forName("com.mysql.cj.jdbc.Driver");
-        } catch (ClassNotFoundException e) {
-            System.err.println(" !! JDBC Driver load 오류: " + e.getMessage());
-            e.printStackTrace();
-        }
-
-        // 드라이버 연결
-        try {
-            con = DriverManager.getConnection("jdbc:mysql://" + server + "/" + database + option, userName, password);
-            System.out.println("정상적으로 연결되었습니다.");
-        } catch (SQLException e) {
-            System.err.println("연결 오류:" + e.getMessage());
-            e.printStackTrace();
-        }
-        return con;
-    }
-}
